refactor(ui): AssetInfo - convert to typescript
