+++ packages/ui/src/components/asset-info/asset-info.stories.tsx
